requirements.txt
langgraph>=0.2.0
langgraph-sdk>=0.1.0
deepagents>=0.1.0
langchain>=0.3.0
langchain-openai>=0.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
fastapi>=0.115.0
apscheduler>=3.10.0
